✨ Add PollModeForm + improvements
