+++ apps/site/src/app/[locale]/(funnel-mode)/organisations/[orgaSlug]/creer-campagne/_components/CloseButton.tsx
@@ -11,8 +11,12 @@ export default function CloseButton({ t }: Props) {
     <Button
       color="secondary"
       aria-label={t(
-        'organisations.createPoll.type.closeButton',
+        'organisations.createPoll.type.closeButton.aria',
         "Abandonner le processus de création de test collectif et revenir à la page d'accueil de votre organisation"
+      )}
+      title={t(
+        'organisations.createPoll.type.closeButton.title',
+        "Abandonner la création du test collectif et revenir à la page d'accueil de votre organisation"
       )}
       className="h-10 w-10 rounded-full p-0!">
       <CloseIcon className="fill-primary-800 max-h-6 min-w-6" />
